+++ admin/inc/logging.class.php
@@ -129,6 +129,4 @@ class GS_Logging_Class {
                 }  
         }
   
-} // end of class                   
-
-?>
+} // end of class