+++ admin/lang/en_US.php
@@ -56,7 +56,8 @@ $i18n = array(
 "MENUITEM_SUBTITLE"	=>	"menu item",
 "HOMEPAGE_SUBTITLE"	=>	"homepage",
 "PRIVATE_SUBTITLE"	=>	"private",
-"COMPONENT_SUBTITLE" => "component enable",
+"COMPONENT_SUBTITLE" => "component enabled",
+"PERMALINK_SUBTITLE" => 'custom permalink',
 "EDITPAGE_TITLE"		=>	"Edit Page",
 "VIEWPAGE_TITLE"		=>	"View Page",
 "DELETEPAGE_TITLE"	=>	"Delete Page",
@@ -222,7 +223,7 @@ $i18n = array(
 "WEBSITE_SETTINGS"	=>	"Website Settings",
 "LOCAL_TIMEZONE"		=>	"Local Timezone",
 "LANGUAGE"					=>	"Language",
-"USE_FANCY_URLS"		=>	"Use Fancy URLs - <b style=\"font-weight:100\">Requires that your host has <code>mod_rewrite</code> enabled</b>",
+"USE_PRETTY_URLS"		=>	"Use Pretty URLs - <b style=\"font-weight:100\">Requires that your host has <code>mod_rewrite</code> enabled</b>",
 "ENABLE_HTML_ED"		=>	"Enable the HTML editor - <b style=\"font-weight:100\">Adds WYSIWYG, code highlights and other features to page content editor</b>",
 "WARN_EMAILINVALID"	=>	"WARNING: This email address does not look valid!",
 "ONLY_NEW_PASSWORD"	=>	"Only provide a password below if you want to change your current one",
@@ -403,8 +404,8 @@ $i18n = array(
 /* 
  * For: support.php
 */
-"WELCOME_MSG"				=>	"Thank you for choosing GetSimple as your content management system!",
-"WELCOME_P"					=>	"GetSimple makes managing a website as simple as possible with its best-in-class user interface. We strive to keep the system easy enough for anyone to use, yet powerful enough for a developer to enable all the features that are needed.</p><p><strong>Some first steps that might be useful:</strong></p>",
+"WELCOME_MSG"				=>	"Thank you for choosing <strong>GetSimple Extended CMS</strong> as your content management system!",
+"WELCOME_P"					=>	"<strong>GetSimple Extended CMS</strong> makes managing a website as simple as possible with its best-in-class user interface. We strive to keep the system easy enough for anyone to use, yet powerful enough for a developer to enable all the features that are needed.</p><p><strong>Some first steps that might be useful:</strong></p>",
 "GETTING_STARTED"		=>	"Getting Started",
 
 /* 
@@ -602,11 +603,9 @@ $i18n = array(
 "ENABLE_CODE_ED"		=>	"Enable the code editor - <b style=\"font-weight:100\">Highlights code syntax and improves usability of editors for components and theme files</b>", // settings.php
 "MAINTENANCE_MODE" => "MAINTENANCE MODE",
 "MAINTENANCE_WARNING" => "Website is working in <strong>Maintenance Mode</strong> and not available for visitors",
-"ALLOW_ACCESS_IN_MAINTENANCE" => "Access to website in Maintenance Mode - <b style=\"font-weight:100\">Allow access to website while Maintenance Mode is enable</b>",
+"ALLOW_ACCESS_IN_MAINTENANCE" => "Access to website in Maintenance Mode - <b style=\"font-weight:100\">Allow access to website while Maintenance Mode is enabled</b>",
 "LABEL_WEBSITELANG" => "Website Language",
 "DISPLAY_WEBSITELANG" => "A code of language of website content (en, de, fr, it, ru, be, ua, etc.). Set per page language in page editor",
 "LABEL_PAGELANG"  => "Language",
 "PLACEHOLDER_LANG" => "en",
 );
-
-?>