+++ admin/settings.php
@@ -281,10 +281,10 @@ get_template('header', cl($SITENAME).' &raquo; '.i18n_r('GENERAL_SETTINGS'));
 			</p>
 		</div>
 		<p class="inline" ><input name="sitemaintenance" id="sitemaintenance" type="checkbox" value="1"<?php if ($SITEMAINTENANCE == 1) { echo ' checked '; }?> />&nbsp;<label for="sitemaintenance" ><?php i18n('MAINTENANCE_ENABLE');?></label></p>
-		<p class="inline" ><input name="prettyurls" id="prettyurls" type="checkbox" value="1" <?php if ($PRETTYURLS == 1) { echo 'checked'; } ?> />&nbsp;<label for="prettyurls" ><?php i18n('USE_FANCY_URLS');?></label></p>
+		<p class="inline" ><input name="prettyurls" id="prettyurls" type="checkbox" value="1" <?php if ($PRETTYURLS == 1) { echo 'checked'; } ?> />&nbsp;<label for="prettyurls" ><?php i18n('USE_PRETTY_URLS');?></label></p>
 				
 		<div class="leftsec">
-			<p><label for="permalink"  class="clearfix"><?php i18n('PERMALINK');?>: <span class="right"><a href="https://github.com/dimayakovlev/getsimple-extended-cms/wiki/Fancy-URLs" target="_blank" ><?php i18n('MORE');?></a></span></label><input class="text" name="permalink" id="permalink" type="text" placeholder="%parent%/%slug%/" value="<?php if(isset($PERMALINK)) { echo var_out($PERMALINK); } ?>" /></p>
+			<p><label for="permalink"  class="clearfix"><?php i18n('PERMALINK');?>: <span class="right"><a href="https://github.com/dimayakovlev/getsimple-extended-cms/wiki/Pretty-URLs" target="_blank" ><?php i18n('MORE');?></a></span></label><input class="text" name="permalink" id="permalink" type="text" placeholder="%parent%/%slug%/" value="<?php if(isset($PERMALINK)) { echo var_out($PERMALINK); } ?>" /></p>
 		<a id="flushcache" class="button" href="?flushcache"><?php i18n('FLUSHCACHE'); ?></a>
 		</div>
 		<div class="clear"></div>