+++ admin/setup.php
@@ -163,7 +163,7 @@ if(isset($_POST['submitted'])) {
 		 	$temp = GSROOTPATH .'temp.htaccess';
 		 	$init = GSROOTPATH.'.htaccess';
 			
-			if(file_exists($temp)) {				
+			if(file_exists($temp)) {
 				$temp_data = file_get_contents(GSROOTPATH .'temp.htaccess');
 				$temp_data = str_replace('**REPLACE**',tsl($path_parts), $temp_data);
 				$fp = fopen($init, 'w');