+++ index.php
@@ -122,8 +122,8 @@ exec_action('index-post-dataindex');
 
 # check for correctly formed url
 if ($GSCANONICAL == true) {
-	if (strpos($_SERVER['REQUEST_URI'], find_url($url, $parent, 'relative')) !== 0) {
-		redirect(find_url($url, $parent));
+	if (strpos($_SERVER['REQUEST_URI'], find_url($url, false)) !== 0) {
+		redirect(find_url($url, true));
 	}
 }